refactor: 알림 설정 관련을 MemberController로 통합 #88
