+++ src/main/java/com/umc/hwaroak/controller/MemberController.java
@@ -1,7 +1,10 @@
 package com.umc.hwaroak.controller;
 
+import com.umc.hwaroak.dto.request.AlarmSettingRequestDto;
+import com.umc.hwaroak.dto.response.AlarmSettingResponseDto;
 import com.umc.hwaroak.dto.response.MemberResponseDto;
 import com.umc.hwaroak.dto.request.MemberRequestDto;
+import com.umc.hwaroak.service.AlarmSettingService;
 import com.umc.hwaroak.service.EmotionSummaryService;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.media.Content;
@@ -22,6 +25,7 @@ public class MemberController {
 
     private final MemberService memberService;
     private final EmotionSummaryService emotionSummaryService;
+    private final AlarmSettingService alarmSettingService;
 
     @GetMapping("")
     @Operation(summary = "회원 정보 조회", description = "회원 정보를 조회합니다.")
@@ -85,4 +89,19 @@ public class MemberController {
         return emotionSummaryService.getDetailEmotionSummary(summaryMonth);
     }
 
+    @GetMapping("/alarmSetting")
+    @Operation(summary = "알림 설정 조회", description = "사용자의 알림 관련 설정들을 조회합니다.")
+    @ApiResponse(content = @Content(schema = @Schema(implementation = AlarmSettingResponseDto.InfoDto.class)))
+    public AlarmSettingResponseDto.InfoDto getAlarmSettingInfo() {
+        return alarmSettingService.getAlarmSettingInfo();
+    }
+
+    @PatchMapping("/alarmSetting")
+    @Operation(summary = "알림 설정 변경", description = "사용자의 알림 관련 설정을 변경합니다. 변경할 필드와 데이터만 넣으면 됩니다.")
+    @ApiResponse(content = @Content(schema = @Schema(implementation = AlarmSettingResponseDto.InfoDto.class)))
+    public AlarmSettingResponseDto.InfoDto editAlarmSettingInfo(
+            @RequestBody AlarmSettingRequestDto.EditDto requestDto
+    ) {
+        return alarmSettingService.editAlarmSettingInfo(requestDto);
+    }
 }
