+++ src/main/java/com/umc/hwaroak/domain/AlarmSetting.java
@@ -2,9 +2,7 @@ package com.umc.hwaroak.domain;
 
 import com.umc.hwaroak.domain.common.BaseEntity;
 import jakarta.persistence.*;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.Setter;
+import lombok.*;
 
 import java.time.LocalDateTime;
 import java.time.LocalTime;
@@ -13,6 +11,8 @@ import java.time.LocalTime;
 @Table(name = "alarm_setting")
 @Getter
 @Builder
+@NoArgsConstructor
+@AllArgsConstructor
 public class AlarmSetting extends BaseEntity {
 
     @Id @GeneratedValue(strategy = GenerationType.IDENTITY)
